requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
executing==2.0.1
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.12.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.1
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
langchain-openai==0.0.2
langchainhub==0.1.14
langsmith==0.0.78
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
openai==1.7.0
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.1.0
posthog==3.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==3.17.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0